+++ lib/src/models/calendar/view_state/multi_day_view_state.dart
@@ -14,7 +14,12 @@ class MultiDayViewState extends ViewState {
     required this.numberOfPages,
     required this.pageController,
     required this.scrollController,
-  });
+  }) {
+    scrollController.addListener(() {
+      _updateTimeOfDay();
+      lastKnowScrollOffset = scrollController.offset;
+    });
+  }
 
   /// Creates a [MultiDayViewState] from a [MultiDayViewConfiguration].
   factory MultiDayViewState.fromViewConfiguration({
@@ -46,14 +51,11 @@ class MultiDayViewState extends ViewState {
       selectedDate,
     );
 
-    final scrollController = ScrollController(
-      keepScrollOffset: true,
-      initialScrollOffset: previousState?.lastKnowScrollOffset ?? 0.0,
-    );
-
-    scrollController.addListener(() {
-      previousState?.lastKnowScrollOffset = scrollController.offset;
-    });
+    final scrollController = previousState?.scrollController ??
+        ScrollController(
+          keepScrollOffset: true,
+          initialScrollOffset: previousState?.lastKnowScrollOffset ?? 0.0,
+        );
 
     return MultiDayViewState(
       visibleDateTimeRange: ValueNotifier(visibleDateRange),
@@ -76,6 +78,11 @@ class MultiDayViewState extends ViewState {
     visibleMonth = visibleDateTimeRangeNotifier.value.start.startOfMonth;
   }
 
+  @override
+  set visibleStartTimeOfDay(TimeOfDay? value) {
+    visibleStartTimeOfDayNotifier.value = value;
+  }
+
   /// The pageController of the current view.
   final PageController pageController;
 
@@ -89,6 +96,9 @@ class MultiDayViewState extends ViewState {
   /// The scrollPhysics of the current view.
   ValueNotifier<ScrollPhysics> scrollPhysics =
       ValueNotifier<ScrollPhysics>(const ScrollPhysics());
+
+  /// The page height set by the widget.
+  double pageHeight = 10;
 
   /// The adjusted dateTimeRange of the current view.
   @override
@@ -108,13 +118,40 @@ class MultiDayViewState extends ViewState {
     Duration? duration,
     Curve? curve,
   }) async {
+    // Calculate the pageNumber of the date.
     final pageNumber = viewConfiguration.calculateDateIndex(
       date,
       adjustedDateTimeRange.start.startOfDay,
     );
 
+    // Animate to that page.
     await pageController.animateToPage(
       pageNumber,
+      duration: duration ?? const Duration(milliseconds: 300),
+      curve: curve ?? Curves.easeInOut,
+    );
+  }
+
+  @override
+  Future<void> animateToDateTime(
+    DateTime date, {
+    Duration? duration,
+    Curve? curve,
+  }) async {
+    await animateToDate(date, duration: duration, curve: curve);
+
+    if (!scrollController.hasClients) return;
+
+    // Calculate the offset of the time.
+    final startTime = date.startOfDay.copyWith(
+      hour: viewConfiguration.startHour,
+    );
+    final timeOffset =
+        date.difference(startTime).inMinutes * heightPerMinute.value;
+
+    // Animate to the offset of the time.
+    await scrollController.animateTo(
+      timeOffset,
       duration: duration ?? const Duration(milliseconds: 300),
       curve: curve ?? Curves.easeInOut,
     );
@@ -129,17 +166,20 @@ class MultiDayViewState extends ViewState {
   }) async {
     // First animate to the date of the event.
     await animateToDate(
-      event.dateTimeRange.start,
+      event.dateTimeRange.start.startOfDay,
       duration: duration ?? const Duration(milliseconds: 300),
       curve: curve ?? Curves.ease,
     );
+
     // Calculate the event position.
+
+    final eventStart = event.start;
+    final startOfDate = eventStart.startOfDay;
+
     final eventPosition =
-        event.start.difference(event.start.startOfDay).inMinutes *
-            heightPerMinute.value;
+        eventStart.difference(startOfDate).inMinutes * heightPerMinute.value;
 
     double scrollPosition;
-
     if (centerEvent) {
       // Calculate the event height.
       final eventHeight = event.duration.inMinutes * heightPerMinute.value;
@@ -201,6 +241,16 @@ class MultiDayViewState extends ViewState {
   @override
   void jumpToPage(int page) {
     pageController.jumpToPage(page);
+  }
+
+  void _updateTimeOfDay() {
+    final total = (scrollController.offset / pageHeight) * 1440;
+    final minute = (total % 60).toInt();
+    final hour = viewConfiguration.startHour + total ~/ 60;
+    visibleStartTimeOfDayNotifier.value = TimeOfDay(
+      hour: hour,
+      minute: minute,
+    );
   }
 
   @override