+++ lib/src/models/calendar/view_state/view_state.dart
@@ -15,6 +15,11 @@ abstract class ViewState {
   DateTimeRange get visibleDateTimeRange => visibleDateTimeRangeNotifier.value;
   set visibleDateTimeRange(DateTimeRange value);
 
+  final ValueNotifier<TimeOfDay?> visibleStartTimeOfDayNotifier =
+      ValueNotifier(null);
+  TimeOfDay? get visibleStartTimeOfDay => visibleStartTimeOfDayNotifier.value;
+  set visibleStartTimeOfDay(TimeOfDay? value) {}
+
   /// The visible month notifier of the current page.
   late DateTime visibleMonth =
       visibleDateTimeRangeNotifier.value.start.startOfMonth;
@@ -22,10 +27,10 @@ abstract class ViewState {
   /// The adjusted dateTimeRange of the current view.
   DateTimeRange get adjustedDateTimeRange;
 
-  /// Jump to the given page.
+  /// Jump to the given [DateTime].
   void jumpToPage(int page);
 
-  /// Jump to the given date.
+  /// Jump to the given [DateTime].
   void jumpToDate(DateTime date);
 
   /// Adjust the height per minute of the current view.
@@ -43,14 +48,21 @@ abstract class ViewState {
     Curve? curve,
   });
 
-  /// Animate to the given date.
+  /// Animate to the given [DateTime].
   Future<void> animateToDate(
     DateTime date, {
     Duration? duration,
     Curve? curve,
   });
 
-  /// Animate to the given event.
+  /// Animate to the given [DateTime].
+  Future<void> animateToDateTime(
+    DateTime date, {
+    Duration? duration,
+    Curve? curve,
+  });
+
+  /// Animate to the given [CalendarEvent].
   Future<void> animateToEvent(
     CalendarEvent event, {
     Duration? duration,