+++ lib/src/models/view_configurations/multi_day_configurations/custom_multi_day_configuration.dart
@@ -24,7 +24,8 @@ class CustomMultiDayConfiguration extends MultiDayViewConfiguration {
     super.startHour = 0,
     super.endHour = 24,
     super.createEventTrigger,
-    super.showHeader,
+    super.showDayHeader,
+    super.showMultiDayHeader,
   });
 
   @override
@@ -42,21 +43,15 @@ class CustomMultiDayConfiguration extends MultiDayViewConfiguration {
     int? firstDayOfWeek,
   }) {
     return DateTimeRange(
-      start: visibleStart.startOfDay.subtract(
-        Duration(
-          days: (visibleStart.difference(dateTimeRange.start).inDays ~/
-                      numberOfDays)
-                  .ceil() *
-              numberOfDays,
-        ),
+      start: visibleStart.startOfDay.subtractDays(
+        (visibleStart.difference(dateTimeRange.start).inDays ~/ numberOfDays)
+                .ceil() *
+            numberOfDays,
       ),
-      end: visibleStart.startOfDay.add(
-        Duration(
-          days: (dateTimeRange.end.difference(visibleStart).inDays ~/
-                      numberOfDays)
-                  .ceil() *
-              numberOfDays,
-        ),
+      end: visibleStart.startOfDay.addDays(
+        (dateTimeRange.end.difference(visibleStart).inDays ~/ numberOfDays)
+                .ceil() *
+            numberOfDays,
       ),
     );
   }